+++ CAESAR/src/run/Run.java
@@ -8,6 +8,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import accident.*;
 import event.*;
 
@@ -546,7 +547,7 @@ public class Run {
 	 * @param distrProgr			distributor progress in event.sec
 	 */
 	public void trafficManagement (PositionReport event, long startOfSimulation, double segWithAccAhead, 
-			AtomicBoolean accidentWarningsFailed, AtomicBoolean tollNotificationsFailed, AtomicInteger max_latency, long distrProgr) {
+			AtomicBoolean accidentWarningsFailed, AtomicBoolean tollNotificationsFailed, AtomicLong max_latency, long distrProgr) {
 		
 		// Set auxiliary variables
 		double next_min = event.min+1;
@@ -688,7 +689,7 @@ public class Run {
 	 * @param distrProgr			distributor progress in event.sec
 	 */
 	public void accidentManagement (PositionReport event, long startOfSimulation, double segWithAccAhead, boolean run_priorization, 
-			AtomicBoolean accidentWarningsFailed, AtomicInteger max_latency, long distrProgr) {
+			AtomicBoolean accidentWarningsFailed, AtomicLong max_latency, long distrProgr) {
 		
 		//System.out.println(event.timesToString());
 		
@@ -768,7 +769,7 @@ public class Run {
 	 * @param distrProgr			distributor progress in event.sec
 	 */
 	public void congestionManagement (PositionReport event, long startOfSimulation, double segWithAccAhead, 
-			AtomicBoolean tollNotificationsFailed, AtomicInteger max_latency, long distrProgr) { 
+			AtomicBoolean tollNotificationsFailed, AtomicLong max_latency, long distrProgr) { 
 		
 		// Set auxiliary variables
 		double next_min = event.min+1;
