+++ CAESAR/src/scheduler/TimeDrivenScheduler.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import run.*;
 import distributor.*;
 import iogenerator.*;
@@ -25,7 +26,7 @@ public class TimeDrivenScheduler extends Scheduler implements Runnable {
 	boolean reduced_stream_history_traversal;
 			
 	public TimeDrivenScheduler (boolean sq, AtomicInteger dp, HashMap<Double,Long> distrProgrPerSec, HashMap<RunID,Run> rs, RunQueues rq, ExecutorService e, 
-			CountDownLatch tn, CountDownLatch d, ArrayList<XwayDirPair> xds, int lastS, long start, AtomicInteger max_latency,
+			CountDownLatch tn, CountDownLatch d, ArrayList<XwayDirPair> xds, int lastS, long start, AtomicLong max_latency,
 			boolean ed, boolean pr, boolean fi, boolean sh) {	
 		
 		super(dp,distrProgrPerSec,rs,rq,e,tn,d,xds,lastS,start,max_latency);
